Prevent swipe back (iOS) and back button (Android) from bypassing the leaveDialog
